Move files all into packages
