+++ src/main/java/cow/exceptions/CowExceptions.java
@@ -1,0 +1,7 @@
+package cow.exceptions;
+
+public class CowExceptions extends Exception  {
+    public CowExceptions(String exception) {
+        super(exception);
+    }
+}
