+++ src/main/java/cow/exceptions/MissingParametersException.java
@@ -1,0 +1,7 @@
+package cow.exceptions;
+
+public class MissingParametersException extends Exception {
+    public MissingParametersException(String command, String expected) {
+        super(command + " has missing parameters! Example: " + expected);
+    }
+}
