+++ src/main/java/cow/exceptions/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package cow.exceptions;
+
+public class UnknownCommandException extends Exception {
+    public UnknownCommandException() {
+        super("Command not found!");
+    }
+}
