--- src/main/java/exceptions/CowExceptions.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class CowExceptions extends Exception  {
-    public CowExceptions(String exception) {
-        super(exception);
-    }
-}
