--- src/main/java/exceptions/InvalidTaskException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class InvalidTaskException extends Exception {
-    public InvalidTaskException() {
-        super("Invalid task Index");
-    }
-}
