--- src/main/java/exceptions/MissingParametersException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class MissingParametersException extends Exception {
-    public MissingParametersException(String command, String expected) {
-        super(command + " has missing parameters! Example: " + expected);
-    }
-}
