--- src/main/java/exceptions/UnknownCommandException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class UnknownCommandException extends Exception {
-    public UnknownCommandException() {
-        super("Command not found!");
-    }
-}
