--- src/main/java/message/Messages.java
@@ -1,8 +1,0 @@
-package message;
-
-public class Messages {
-    public static final String MESSAGE_INVALID_COMMAND_FORMAT = "Invalid command format! \n%s";
-    public static final String INDEX_INVALID = "Invalid index format!";
-    public static final String INDEX_OUT_OF_RANGE = "Invalid index!";
-    public static final String FILE_ERROR = "Unable to write to file!";
-}
