+++ text-ui-test/data/cow.txt
@@ -1,0 +1,3 @@
+T | 0 | borrow book
+D | 0 | return book | 2/12/2019 1800
+E | 0 | project meeting | Mon 2pm | 4pm
